third_party/blink/web_tests/editing/pasteboard/newlines-around-floating-or-positioned.html

<!doctype>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="test1">Lorem<span style="float: left;">-</span>ipsum</div>
<div id="test2">Lorem<span style="float: left;"></span>ipsum</div>
<div id="test3">Lorem<div style="position: absolute;">-</div>ipsum</div>
<div id="test4">Lorem<div style="position: absolute;"></div>ipsum</div>
<script>
test(() => assert_equals(
  document.getElementById('test1').innerText,
  'Lorem\n-\nipsum'),
  '1 float:left minus');
test(() => assert_equals(
  document.getElementById('test2').innerText,
  'Lorem\nipsum'),
  '2 float:left empty');
test(() => assert_equals(
  document.getElementById('test3').innerText,
  'Lorem\n-\nipsum'),
  '3 position:absolute minus');
test(() => assert_equals(
  document.getElementById('test4').innerText,
  'Lorem\nipsum'),
  '4 position:absolute empty');
</script>
